backend/requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
flask-jwt-extended==4.4.4
python-dotenv==1.0.1
pymongo[srv]==4.4.0
flask-cors==4.0.0
cloudinary==1.32.0
werkzeug==2.3.7
pytest==7.4.2
